+++ Flipper/Packages/UI/Sources/Welcome/WelcomeViewModel.swift
@@ -1,0 +1,5 @@
+import Combine
+
+class WelcomeViewModel: ObservableObject {
+    init() {}
+}
